Split remote project integration tests from workspace model
